Removed old wrapper around AlertDialog.Builder, which functions were not used actually.
