+++ app/src/main/java/co/smartreceipts/android/di/GlobalBindingModule.java
@@ -31,10 +31,10 @@ import co.smartreceipts.android.fragments.SelectAutomaticBackupProviderDialogFra
 import co.smartreceipts.android.ocr.widget.di.OcrConfigurationFragmentSubcomponent;
 import co.smartreceipts.android.rating.FeedbackDialogFragment;
 import co.smartreceipts.android.rating.RatingDialogFragment;
-import co.smartreceipts.android.settings.widget.CSVColumnsListFragment;
-import co.smartreceipts.android.settings.widget.CategoriesListFragment;
-import co.smartreceipts.android.settings.widget.PDFColumnsListFragment;
-import co.smartreceipts.android.settings.widget.PaymentMethodsListFragment;
+import co.smartreceipts.android.settings.widget.editors.CSVColumnsListFragment;
+import co.smartreceipts.android.settings.widget.editors.CategoriesListFragment;
+import co.smartreceipts.android.settings.widget.editors.PDFColumnsListFragment;
+import co.smartreceipts.android.settings.widget.editors.PaymentMethodsListFragment;
 import co.smartreceipts.android.settings.widget.SettingsActivity;
 import co.smartreceipts.android.sync.drive.services.DriveCompletionEventService;
 import co.smartreceipts.android.sync.drive.services.di.DriveCompletionEventServiceSubcomponent;
