+++ app/src/main/java/co/smartreceipts/android/di/subcomponents/CSVColumnsListFragmentSubcomponent.java
@@ -1,7 +1,7 @@
 package co.smartreceipts.android.di.subcomponents;
 
 import co.smartreceipts.android.di.scopes.FragmentScope;
-import co.smartreceipts.android.settings.widget.CSVColumnsListFragment;
+import co.smartreceipts.android.settings.widget.editors.CSVColumnsListFragment;
 import dagger.Subcomponent;
 import dagger.android.AndroidInjector;
