+++ app/src/main/java/co/smartreceipts/android/di/subcomponents/CategoriesListFragmentSubcomponent.java
@@ -1,7 +1,7 @@
 package co.smartreceipts.android.di.subcomponents;
 
 import co.smartreceipts.android.di.scopes.FragmentScope;
-import co.smartreceipts.android.settings.widget.CategoriesListFragment;
+import co.smartreceipts.android.settings.widget.editors.CategoriesListFragment;
 import dagger.Subcomponent;
 import dagger.android.AndroidInjector;
