+++ app/src/main/java/co/smartreceipts/android/settings/widget/SettingsViewerActivity.java
@@ -5,6 +5,10 @@ import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 
 import co.smartreceipts.android.R;
+import co.smartreceipts.android.settings.widget.editors.CSVColumnsListFragment;
+import co.smartreceipts.android.settings.widget.editors.CategoriesListFragment;
+import co.smartreceipts.android.settings.widget.editors.PDFColumnsListFragment;
+import co.smartreceipts.android.settings.widget.editors.PaymentMethodsListFragment;
 
 public class SettingsViewerActivity extends AppCompatActivity {
