+++ app/src/main/java/co/smartreceipts/android/settings/widget/editors/PDFColumnsListFragment.java
@@ -1,4 +1,4 @@
-package co.smartreceipts.android.settings.widget;
+package co.smartreceipts.android.settings.widget.editors;
 
 import android.content.Context;
 import android.support.v7.app.ActionBar;
@@ -7,26 +7,27 @@ import javax.inject.Inject;
 
 import co.smartreceipts.android.R;
 import co.smartreceipts.android.model.impl.columns.receipts.ReceiptColumnDefinitions;
-import co.smartreceipts.android.persistence.database.controllers.impl.CSVTableController;
 import co.smartreceipts.android.persistence.database.controllers.impl.ColumnTableController;
+import co.smartreceipts.android.persistence.database.controllers.impl.PDFTableController;
 import dagger.android.support.AndroidSupportInjection;
 
-public class CSVColumnsListFragment extends ColumnsListFragment {
+public class PDFColumnsListFragment extends ColumnsListFragment {
 
-	public static String TAG = "CSVColumnsListFragment";
+    public static String TAG = "PDFColumnsListFragment";
+
 
     @Inject
     ReceiptColumnDefinitions receiptColumnDefinitions;
     @Inject
-    CSVTableController csvTableController;
+    PDFTableController pdfTableController;
 
-	public static ColumnsListFragment newInstance() {
-		return new CSVColumnsListFragment();
-	}
+    public static ColumnsListFragment newInstance() {
+        return new PDFColumnsListFragment();
+    }
 
     @Override
     protected ColumnTableController getColumnTableController() {
-        return csvTableController;
+        return pdfTableController;
     }
 
     @Override
@@ -41,12 +42,12 @@ public class CSVColumnsListFragment extends ColumnsListFragment {
     }
 
     @Override
-	public void onResume() {
-		super.onResume();
+    public void onResume() {
+        super.onResume();
         final ActionBar actionBar = getSupportActionBar();
         if (actionBar != null) {
-            actionBar.setTitle(R.string.menu_main_csv);
+            getSupportActionBar().setTitle(R.string.menu_main_pdf);
         }
-	}
+    }
 
 }
