+++ app/src/main/java/co/smartreceipts/android/settings/widget/editors/PaymentMethodsListFragment.java
@@ -1,4 +1,4 @@
-package co.smartreceipts.android.settings.widget;
+package co.smartreceipts.android.settings.widget.editors;
 
 import android.app.AlertDialog;
 import android.content.Context;
