+++ app/src/main/java/co/smartreceipts/android/trips/TripFragment.java
@@ -45,7 +45,6 @@ import co.smartreceipts.android.utils.log.Logger;
 import co.smartreceipts.android.widget.tooltip.Tooltip;
 import co.smartreceipts.android.workers.EmailAssistant;
 import dagger.android.support.AndroidSupportInjection;
-import wb.android.dialog.BetterDialogBuilder;
 import wb.android.flex.Flex;
 
 public class TripFragment extends WBListFragment implements TableEventsListener<Trip>, AdapterView.OnItemLongClickListener {
@@ -180,7 +179,7 @@ public class TripFragment extends WBListFragment implements TableEventsListener<
     }
 
     public final boolean editTrip(final Trip trip) {
-        final BetterDialogBuilder builder = new BetterDialogBuilder(getActivity());
+        final AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
         final String[] editTripItems = flex.getStringArray(getActivity(), R.array.EDIT_TRIP_ITEMS);
         builder.setTitle(trip.getName())
                 .setCancelable(true)
@@ -206,7 +205,7 @@ public class TripFragment extends WBListFragment implements TableEventsListener<
     }
 
     public final void deleteTrip(final Trip trip) {
-        final BetterDialogBuilder builder = new BetterDialogBuilder(getActivity());
+        final AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
         builder.setTitle(getString(R.string.delete_item, trip.getName()))
                 .setMessage(getString(R.string.delete_sync_information))
                 .setCancelable(true)
