+++ gradle.properties
@@ -37,6 +37,8 @@ PDFIUM_VERSION=1.5.0
 
 MP_ANDROID_CHART_VERSION=v3.0.2
 
+ADV_RECYCLER_VIEW_VERSION=0.10.6
+
 DAGGER_VERSION=2.11-rc2
 
 org.gradle.jvmargs=-Xmx3072m
